fix(SPRING-11): order.price float

+++ src/main/java/com/alsab/boozycalc/dto/OrderDto.java
@@ -12,8 +12,7 @@ import lombok.NoArgsConstructor;
 public class OrderDto {
     private Long id;
     private String name;
-    private Long price;
+    private float price;
     private PartyEntity party_id;
     private UserEntity person_id;
-
 }
